[Fix] 따릉이 관련 뱃지 로직 수정

+++ greenmap/src/main/java/com/ucamp/greenmap/badge/dto/response/BadgeResponse.java
@@ -17,7 +17,7 @@ public class BadgeResponse {
     @Builder
     public static class BadgeInfo {
         private String name;
-        private Long progress;
+        private Double progress;
         private Long standard;
         private String description;
         private String image_url;
